+++ src/pyc_wheel/__about__.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2019-2021 Adam Karpierz
-# Licensed under the MIT License
-# https://opensource.org/licenses/MIT
+# Copyright (c) 2016 Grant Patten
+# Copyright (c) 2019 Adam Karpierz
+# SPDX-License-Identifier: MIT
 
 __import__("pkg_about").about()
-__copyright__ = f"Copyright (c) 2019-2021 {__author__}"  # noqa