+++ src/pyc_wheel/__init__.py
@@ -1,6 +1,7 @@
-# Copyright (c) 2019-2021 Adam Karpierz
-# Licensed under the MIT License
-# https://opensource.org/licenses/MIT
+# Copyright (c) 2016 Grant Patten
+# Copyright (c) 2019 Adam Karpierz
+# SPDX-License-Identifier: MIT
 
-from .__about__  import * ; del __about__   # noqa
+from .__about__ import * ; del __about__  # noqa
+
 from ._pyc_wheel import * ; del _pyc_wheel  # noqa