+++ src/pyc_wheel/__main__.py
@@ -1,6 +1,6 @@
-# Copyright (c) 2019-2021 Adam Karpierz
-# Licensed under the MIT License
-# https://opensource.org/licenses/MIT
+# Copyright (c) 2016 Grant Patten
+# Copyright (c) 2019 Adam Karpierz
+# SPDX-License-Identifier: MIT
 
 import sys
 from ._pyc_wheel import main